inode: add idmapped mount aware init and permission helpers
The inode_owner_or_capable() helper determines whether the caller is the
owner of the inode or is capable with respect to that inode. Allow it to handle
idmapped mounts. If the If the inode is accessed through an idmapped mount we
first need to map it according to the mount's user namespace. Afterwards the
checks are identical to non-idmapped mounts. If the initial user namespace is
passed all operations are a nop so non-idmapped mounts will not see a change in
behavior and will also not see any performance impact.

Similarly, we allow the inode_init_owner() helper to handle idmapped mounts. It
initializes a new inode on idmapped mounts by mapping the fsuid and fsgid of
the caller from the mount's user namespace. If the initial user namespace is
passed all operations are a nop so non-idmapped mounts will not see a change in
behavior and will also not see any performance impact.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>
---
/* v2 */
- Christoph Hellwig:
  - Don't pollute the vfs with additional helpers simply extend the existing
    helpers with an additional argument and switch all callers.