namei: prepare for idmapped mounts
The various vfs_*() helpers are called by filesystems or by the vfs itself to
perform core operations create, link, mkdir, mknod, rename, rmdir, tmpfile and
unlink. Enable them to handle idmapped mounts. If the inode is accessed through
an idmapped mount it is mapped according to the mount's user namespace.
Afterwards the checks and operations are identical to non-idmapped mounts. If
the initial user namespace is passed all mapping operations are a nop so
non-idmapped mounts will not see a change in behavior and will also not see any
performance impact.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>
---
/* v2 */
- Christoph Hellwig:
  - Don't pollute the vfs with additional helpers simply extend the existing
    helpers with an additional argument and switch all callers.