stat: handle idmapped mounts
The generic_fillattr() helper fills in the basic attributes associated with an
inode. Enable it to handle idmapped mounts. If the inode is accessed through an
idmapped mount we need to map it according to the mount's user namespace. If
the initial user namespace is passed all operations are a nop so non-idmapped
mounts will not see a change in behavior and will also not see any performance
impact.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>
---
/* v2 */
- Christoph Hellwig:
  - Don't pollute the vfs with additional helpers simply extend the existing
    helpers with an additional argument and switch all callers.