fs: use vfs_{get,set}_caps() for capabilities xattrs
Now that all the plumbing is in place, switch over to using the new
inode operations to get/set fs caps. This pushes all mapping of ids into
the caller's user ns to above the vfs_*() level, making this consistent
with other vfs_*() interfaces.

With this change, we need to eliminate cap_inode_getsecurity() from
xattr call stacks in this commit to remove any mapping of rootid in the
raw xattr data, as the new helpers want the fs caps xattrs unmodified
from disk. cap_convert_nscap() is updated to work on vfs_caps and moved
to the new set code path.

Signed-off-by: Seth Forshee <sforshee@digitalocean.com>