acl: remove posix acl handlers
Now that everything in the VFS and all filesystems just rely on the
dedicated inode methods to interact with posix acls we can remove the
dummy generic posix acl xattr handlers. This way no filesystem an be
falsely under the impression that they are needed to implement posix acl
support.

Signed-off-by: Christian Brauner (Microsoft) <brauner@kernel.org>