fs: add generic "delegate" mount option
Currently filesystems can mark themselves mountable inside of containers
by raising FS_USERNS_MOUNT. For any distribution that allows the
creation of user namespaces by unprivileged users this is equivalent to
allowing unprvileged users to mount such filesystems.

Making a filesystem mountable in containers thus gives unprivileged
users the ability to create a superblock for the filesystem. For
example, tmpfs is easily mountable by unprivileged users:

        unshare --user --map-root --mount
        mount -t tmpfs tmpfs /mnt

Mapping this to the new mount api the following happens:

        // Inside of an unprivileged sandbox created e.g, via:
        // unshare --user --map-root --mount
        fd_fs = fsopen("tmpfs");

        // create new superblock
        fsconfig(fd_fs, FSCONFIG_CMD_CREATE, NULL, NULL, 0);

        // create mount for superblock
        fd_mnt = fsmount(fd_fs, 0);

        // make mount visible inside of a mount namespace
        move_mount(fd_mnt, "", -EBADF, "/mnt", ...)

For filesystems without an on-disk image format or backing devices the
creation of a superblock in is not necessarily a security sensitive
step. Though even fur such filesystems bugs such as ones found in
overlayfs have immediate widespread and severe security implications.

For filesystems mountable by container such as overlayfs or tmpfs the
namespace of the process that initially created the fs context will
become the owning namespace of the filesystem's superblock.

Filesystems with an on-disk format and that interact with backing
devices do not support superblock creation inside of containers. For
such filesystems the initial namespace will become the owning namespace
of the filesystem's superblock. This is a sound restriction and it has
been clarified repeatedly that this will not be officially supported.

Not to worry though, userspace supports delegated mounting of
filesystems not mountable inside of containers. Thanks to the new mount
api it is possible to put a sufficiently privileged userspace process in
charge of establishing trust and thus to vouch for the safety of the
image and the backing device. This can be achieved in a variety of ways.
For example, by using dm-verity, dm-integrity, trusted paths where
trusted images must reside, signature verification via the kernel
keyring etc.

A first implementation of the protocol is available in the form of
systemd-mountfsd. This is aready partially integrated with mount(8) via
the discoverable disk images (DDI) specification. Such images can be
mounted using the mount -t ddi switch.

In this manner an unprivileged process inside of the container can
already mount images. The way this works behind the scene is that
systemd creates a mount:

        // systemd-mountfds gets mount request from container via
        // AF_UNIX socket and verifies the requested image and
        // creates a superblock:
        fd_fs = fsopen("ext4");
        fsconfig(fd_fs, FSCONFIG_SET_STRING, "source", "/dev/sda", 0);

        // systemd-mountfsd creates superblock
        fsconfig(fd_fs, FSCONFIG_CMD_CREATE, NULL, NULL, 0);

        // systemd-mountfsd creates initial mount for the superblock
        fd_mnt = fsmount(fd_fs, 0);

        // systemd-mountfsd idmaps fd_mnt for the container
        mount_setattr(fd_mnt, /* create idmapped mount */;

        // The resulting detached mount fd_mnt is sent to the container
        // via AF_UNIX socket and the container attaches the mount to
        // its mount namespace.
        move_mount(fd_mnt, "", -EBAD, "/mnt", ...);

In this model the mount attached to the filesystem of the container will
be fully under its control. Since it is an idmapped mount the ownership
will also be under control of the container.

However, crucially, the superblock for such a filesystem is not under
the control of the container. In other words, the container cannot
change mount options of the superblock. For example, it's not possible
for the container to transition the whole filesystem between rw and ro.

Consequently, it's also not possible to create idmapped mounts of the
filesystem that was delegated to the container because the kernel
correctly requires that the container must own the superblock before
allowing the creation of idmapped mounts.

Such use-cases where it is desirable that the container can change
superblock mount options or create idmapped mounts of a filesystem it
was given should be supported without requiring the filesystem to allow
the creation of superblocks by untrusted users.

To this end a the generic "delegate" mount property is introduced via
fsconfig(). This mount property does only make sense in the new mount
api that allows delegation by splitting the creation of fs context,
superblock, and mount into distinct steps.

When the "delegate" mount property is specified the namespace specified
trough a file descriptor will become the owning namespace of the superblock:

        // systemd-mountfds gets mount request from container via
        // AF_UNIX socket and verifies the requested image and
        // creates a superblock:
        fd_fs = fsopen("ext4");
        fsconfig(fd_fs, FSCONFIG_SET_STRING, "source", "/dev/sda", 0);

        // systemd-mountfsd delegates the superblock to fd_namespace
        // once it is created. systemd-mountfsd must be privileged over
        // fd_namespace.
        fsconfig(fd_fs, FSCONFIG_SET_FD, "delegate", NULL, fd_namespace);

        // systemd-mountfsd creates superblock and delegates it to
        // fd_namespace.
        fsconfig(fd_fs, FSCONFIG_CMD_CREATE, NULL, NULL, 0);

        // systemd-mountfsd creates initial mount for the superblock
        fd_mnt = fsmount(fd_fs, 0);

        // The resulting detached mount fd_mnt is sent to the container
        // via AF_UNIX socket and the container attaches the mount to
        // its mount namespace.
        move_mount(fd_mnt, "", -EBAD, "/mnt", ...);

Afterwards the container owns the superblock and is free to change
superblock mount options or create idmapped mounts. By default, any
filesystem that supports idmapped mounts via FS_ALLOW_IDMAP is eligible
for superblock delegation. Should it be of interest to filesystems
without idmapped mount support they can be supported as well.

Any attempt to create the superblock for a filesystem not mountable
inside of container will fail inside of a container. However, a process
privileged in the initial namespace is able to create a delegated
superblock that can be handed off to a container as illustrated above.

Note, as with any api, the mount api requires discipline when used by
userspace. A privileged process should never interact with an untrusted
filesystem context file descriptor. All security sensitive steps should
be taken in a trusted context. Untrusted filesystem context file
descriptors should be discarded. This situation will be improved once
the mount api grows introspection abilities.

Signed-off-by: Christian Brauner <brauner@kernel.org>