fs: introduce fs_supers_lock
Currently we use a global spin lock for all filesystems to protect the
fs_type->fs_supers list a few other things even though there's no clear
need to do this. This means that everytime a superblock for a given
filesystem type is allocated we hold a global spin lock. When you have
nfs mounts on the system where you may get a large number of superblocks
allocted due to automounts this becomes potentially costly. Also,
userspace can use the ustat(2) system call to walk that list holding
sb lock. Introduce a per-filesystem type spin lock that only protects
super block instances of that filesystem. We'll use that in the next
patch to decouple the global super blocks list and superblock
allocations.

If there's a better way to initialize the per filesystem spinlock I'm
all ears but keep in mind that we also need to capture all filesystems
that aren't registered. IOW, it's not possible to initialize the spin
lock during register_filesystem() as that would miss quite a few
filesystems.

Signed-off-by: Christian Brauner <brauner@kernel.org>